Implement the ProcessPleOp function
CascadingCompiler generates the cascading command stream by processing
each Op in the MergedOpGraph. This function processes the Ple Ops, adds
the corresponding agents to the command stream, and adds the required
dependencies between the agents.

+++ driver/support_library/command_stream/include/ethosn_command_stream/cascading/CommandStream.hpp
@@ -348,8 +348,8 @@ struct AgentData
 /// that are needed by other/this agent
 struct Ratio
 {
-    uint8_t other;
-    uint8_t self;
+    uint16_t other;
+    uint16_t self;
 };
 
 /// Used to represent a dependency between this agent and some other agent
